+++ jwt-util/jwt_util/go/util/str_util.go
@@ -1,0 +1,13 @@
+package util
+
+func strptr(data string) *string {
+	return &data
+}
+
+func strenc(data string) []byte {
+	return ByteArray(data)
+}
+
+func B64EncStr(data string) string {
+	return b64enc(strenc(data))
+}
