+++ jwt-util/jwt_util/go/util/type_alias.go
@@ -1,0 +1,4 @@
+package util
+
+type ByteArray []byte
+type StringMap map[string]any
